投稿例を各箱ごとに変更
