+++ app/assets/stylesheets/custom.css
@@ -29,6 +29,10 @@ body {
 .header-logo:hover { transform: scale(1.07); text-shadow: 2px 2px 0 #fff, 4px 4px 10px rgba(0,0,0,.25); }
 .header-subtitle { font-size: .9rem; color: #666; margin-left: 8px; }
 
+.header-logo {
+  text-decoration: none;
+}
+
 /* ---- ナビ ---- */
 .header-nav {
   display: flex;
@@ -210,7 +214,9 @@ body {
 .modal {
   position:fixed; inset:0; z-index:100;
   display:flex; justify-content:center; align-items:flex-start; overflow-y:auto;
-  background: rgba(255, 230, 225, 0.9); backdrop-filter:blur(10px);
+  background: rgba(255, 230, 225, 0.9);
+  -webkit-backdrop-filter: blur(10px);
+  backdrop-filter: blur(10px);
   opacity:0; pointer-events:none; transition:opacity .4s; padding:60px 0;
 }
 .modal.active { opacity:1; pointer-events:all; }
@@ -267,6 +273,15 @@ body {
 .pink-box .example-box {
   background:#ffdbe2; border:1px solid #ff8fa2; color:#8b1a2b;
 }
+
+.green-box ul,
+.purple-box ul,
+.pink-box ul {
+  list-style: none;
+  padding-left: 0;
+  margin-left: 0;
+}
+
 
 /* ---- ログインボタン ---- */
 .login-button {
@@ -369,4 +384,3 @@ body.winter .mypage-title { color: #60a5fa; }
   letter-spacing: 0.02em;
   transition: color 0.3s ease;
 }
-
