+++ app/assets/stylesheets/post_form.css
@@ -1,6 +1,3 @@
-/* ===============================
-   投稿タイプカード
-================================ */
 .post-type-card {
   position: relative;
   border: 3px solid #e5e7eb;
@@ -82,9 +79,6 @@
   transform: translateY(-5px);
 }
 
-/* ===============================
-   意見カード
-================================ */
 .opinion-radio:checked ~ .opinion-content {
   border-color: #ff9a8e !important;
   background: linear-gradient(135deg, #fff0eb 0%, #ffe4db 100%);
