+++ app/views/posts/_actions.html.erb
@@ -1,0 +1,23 @@
+<% if user_signed_in? %>
+  <div class="post-actions">
+    <!-- 花ボタン -->
+    <button
+      class="flower-btn action-icon"
+      data-post-id="<%= post.id %>"
+    >
+      🌱 <span class="like-count hidden">0</span>
+    </button>
+
+    <!-- コメントボタン -->
+    <button
+      class="action-icon"
+      data-action="click->comment#open"
+    >
+      💬 <span class="comment-count hidden">0</span>
+    </button>
+  </div>
+<% else %>
+  <div class="login-hint">
+    💡 ログインすると「いいね」や「コメント」ができます
+  </div>
+<% end %>
