+++ app/views/posts/index.html.erb
@@ -30,9 +30,14 @@
     <!-- 投稿グリッド -->
     <div id="posts-grid" class="grid md:grid-cols-2 lg:grid-cols-3 gap-6">
       <% @posts.each do |post| %>
-        <div class="post-card"
-             data-category="<%= post.post_type %>"
-             data-created-at="<%= post.created_at.iso8601 %>">
+        <%= link_to post_path(post),
+          class: "post-card block no-underline hover:shadow-xl transition transform hover:-translate-y-1",
+          data: {
+            turbo_frame: "_top",
+            id: post.id,
+            created_at: post.created_at.iso8601,
+            category: post.post_type
+          } do %>
 
           <!-- 投稿ヘッダー -->
           <div class="post-card-header <%= post.post_type %>">
@@ -41,7 +46,7 @@
                  when 'future' then %>🌱 未来宣言箱
                  <% when 'organize' then %>🌈 心の整理箱
                  <% when 'thanks' then %>💌 感謝箱
-                 <% else %>📦 エラー
+                 <% else %>📦 不明
                  <% end %>
             </span>
             <span class="post-date"><%= l post.created_at, format: :short %></span>
@@ -56,13 +61,13 @@
             <% end %>
 
             <p class="post-content text-gray-700 leading-relaxed">
-              <%= simple_format(h(post.body)) %>
+              <%= truncate(strip_tags(post.body), length: 100) %>
             </p>
           </div>
 
           <!-- 投稿フッター -->
           <div class="post-card-footer">
-            <div class="post-meta">
+            <div class="post-meta mb-2">
               <% if post.is_anonymous? %>
                 <span class="post-author">匿名さん</span>
               <% else %>
@@ -74,14 +79,11 @@
               <% end %>
             </div>
 
-            <div class="post-actions">
-              <% if post.opinion_needed? %>
-                <button class="action-icon comment-btn">💬</button>
-              <% end %>
-              <button class="action-icon flower-btn">🌱</button>
-            </div>
+            <!-- ✅ 共通アクション読込 -->
+            <%= render "posts/actions", post: post %>
           </div>
-        </div>
+
+        <% end %>
       <% end %>
 
       <% if @posts.empty? %>
@@ -101,4 +103,4 @@
   </div>
 </main>
 
-<%= javascript_include_tag "posts", "data-turbo-track": "reload", defer: true %>
+<%= javascript_import_module_tag "posts" %>
