+++ app/views/posts/show.html.erb
@@ -1,0 +1,83 @@
+<main class="flex-grow pt-24 pb-12">
+  <div class="container mx-auto max-w-4xl px-6">
+    
+    <!-- 戻るボタン -->
+    <div class="mb-6">
+      <%= link_to "← 投稿一覧に戻る", posts_path, class: "text-gray-600 hover:text-gray-800 transition" %>
+    </div>
+
+    <!-- 投稿カード -->
+    <div class="post-detail-card bg-white rounded-3xl shadow-lg overflow-hidden">
+      
+      <!-- ヘッダー -->
+      <div class="post-header <%= @post.post_type_color %> p-6">
+        <div class="flex items-center justify-between">
+          <div class="flex items-center gap-3">
+            <span class="text-4xl"><%= @post.post_type_icon %></span>
+            <div>
+              <h2 class="text-2xl font-bold text-gray-800"><%= @post.post_type_name %></h2>
+              <p class="text-sm text-gray-600"><%= @post.created_at.strftime("%Y年%m月%d日 %H:%M") %></p>
+            </div>
+          </div>
+          
+          <!-- 投稿者のみ：編集・削除ボタン -->
+          <% if user_signed_in? && @post.user == current_user %>
+            <div class="flex gap-2">
+              <%= link_to edit_post_path(@post), class: "px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600 transition" do %>
+                ✏️ 編集
+              <% end %>
+              <%= button_to @post, method: :delete, 
+                  data: { turbo_confirm: "本当に削除しますか？この操作は取り消せません。" },
+                  class: "px-4 py-2 bg-red-500 text-white rounded-lg hover:bg-red-600 transition" do %>
+                🗑️ 削除
+              <% end %>
+            </div>
+          <% end %>
+        </div>
+
+        <!-- 意見募集ステータス（心の整理箱のみ） -->
+        <% if @post.organize? %>
+          <div class="mt-4">
+            <% if @post.opinion_needed %>
+              <span class="inline-block px-4 py-2 bg-blue-100 text-blue-700 rounded-full text-sm font-bold">
+                💬 意見募集中
+              </span>
+            <% else %>
+              <span class="inline-block px-4 py-2 bg-gray-100 text-gray-700 rounded-full text-sm font-bold">
+                💭 意見不要
+              </span>
+            <% end %>
+          </div>
+        <% end %>
+      </div>
+
+      <!-- 本文 -->
+      <div class="post-body p-8">
+        <div class="prose max-w-none">
+          <p class="text-lg text-gray-800 leading-relaxed whitespace-pre-wrap"><%= @post.body %></p>
+        </div>
+      </div>
+
+      <!-- フッター -->
+      <div class="post-footer p-6 bg-gray-50 border-t">
+        <div class="flex items-center justify-between">
+          <div class="flex items-center gap-4">
+            <span class="text-gray-700">
+              <span class="font-bold">投稿者:</span> <%= @post.display_name %>
+            </span>
+          </div>
+
+          <!-- ✅ 共通アクション -->
+          <%= render "posts/actions", post: @post %>
+        </div>
+      </div>
+    </div>
+
+    <!-- コメント欄（将来実装） -->
+    <div class="mt-8 bg-white rounded-3xl shadow-lg p-8">
+      <h3 class="text-2xl font-bold text-gray-800 mb-6">💬 コメント</h3>
+      <p class="text-gray-500 text-center py-8">コメント機能はMVP後に実装予定です</p>
+    </div>
+
+  </div>
+</main>
