+++ app/views/static_pages/top.html.erb
@@ -18,7 +18,11 @@
         </p>
       </div>
 
-      <a href="#" class="btn-primary px-8 py-4 text-xl rounded-lg">今すぐはじめる →</a>
+      <% if user_signed_in? %>
+        <%= link_to "マイページへ →", mypage_path, class: "btn-primary px-8 py-4 text-xl rounded-lg" %>
+      <% else %>
+        <%= link_to "ログインしてはじめる →", new_user_session_path, class: "btn-primary px-8 py-4 text-xl rounded-lg" %>
+      <% end %>
     </div>
   </section>
 </main>
