+++ config/routes.rb
@@ -1,17 +1,12 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  # Deviseユーザー認証
   devise_for :users
-
-  # マイページ関連
-  get 'mypage', to: 'users#mypage', as: :mypage
-  get 'change_season', to: 'users#change_season', as: :change_season
-
-  # 投稿関連
-  resources :posts, only: [:index, :show, :new, :create]
-
-  # トップページ
   root 'static_pages#top'
 
+  # 投稿関連（全アクション）
+  resources :posts, only: [:index, :show, :new, :create, :edit, :update, :destroy]
+
+  # マイページ
+  get 'mypage', to: 'users#mypage', as: :mypage
 end
